Close atomic upload residual-data gaps for 1.0.7
