--- .github/workflows/one-shot-107-hardening.yml
@@ -1,77 +1,0 @@
-name: Apply Ghost FTP 1.0.7 residual-upload hardening
-
-on:
-  push:
-    branches:
-      - release/ghost-ftp-1.0.7-quality
-    paths:
-      - scripts/one_shot_107_hardening.py
-      - .github/workflows/one-shot-107-hardening.yml
-
-permissions:
-  contents: write
-
-jobs:
-  patch:
-    runs-on: ubuntu-24.04
-    steps:
-      - name: Checkout patch branch
-        uses: actions/checkout@3d3c42e5aac5ba805825da76410c181273ba90b1
-        with:
-          ref: release/ghost-ftp-1.0.7-quality
-          fetch-depth: 0
-
-      - name: Apply deterministic patch
-        run: python3 scripts/one_shot_107_hardening.py
-
-      - name: Remove and stage one-shot tooling before production audits
-        shell: bash
-        run: |
-          set -euo pipefail
-          rm -f scripts/one_shot_107_hardening.py
-          find .github/workflows -maxdepth 1 -type f -name 'one-shot-*' -print -delete
-          git add -A
-          if git ls-files '.github/workflows/one-shot-*' | grep -q .; then
-            echo 'Tracked one-shot workflow remains after staged cleanup.' >&2
-            exit 1
-          fi
-
-      - name: Validate residual cleanup and complete source tree
-        shell: bash
-        run: |
-          set -euo pipefail
-          find 'GhostFTP WEB' -name '*.php' -type f -print0 | xargs -0 -n1 php -l >/tmp/php-lint.log
-          cat /tmp/php-lint.log
-          php 'GhostFTP WEB/tests/atomic-upload-recovery.php'
-          php 'GhostFTP WEB/tests/atomic-upload-residual-cleanup.php'
-          php 'GhostFTP WEB/tests/public-error-boundary.php'
-          php 'GhostFTP WEB/tests/html-error-and-interface-boundary.php'
-          php 'GhostFTP WEB/tests/unit.php'
-          python3 -m unittest discover -s scripts -p 'test_*.py'
-          go test ./...
-          go vet ./...
-          python3 scripts/audit_repository.py
-          python3 scripts/audit_security.py
-          python3 scripts/audit_privacy.py
-          python3 scripts/audit_web.py
-          python3 scripts/audit_version.py
-          python3 scripts/audit_docs.py
-          python3 scripts/audit_release.py
-          git diff --check
-          git diff --cached --check
-
-      - name: Commit clean validated production tree
-        shell: bash
-        run: |
-          set -euo pipefail
-          git config user.name 'Brendigo'
-          git config user.email 'info@brendigo.com'
-          git add -A
-          if git ls-files '.github/workflows/one-shot-*' | grep -q .; then
-            echo 'Tracked one-shot workflow remains after cleanup.' >&2
-            exit 1
-          fi
-          git diff --cached --check
-          git status --short
-          git commit -m 'Close atomic upload residual-data gaps for 1.0.7'
-          git push origin HEAD:release/ghost-ftp-1.0.7-quality
