+++ GhostFTP WEB/VERSION
@@ -1,1 +1,1 @@
-1.0.6
+1.0.7
